+++ tiger-test-lib/src/main/java/de/gematik/test/tiger/lib/rbel/RbelMessageValidator.java
@@ -87,7 +87,7 @@ public class RbelMessageValidator {
         return new UnmodifiableList<>(new ArrayList<>(LocalProxyRbelMessageListener.getValidatableRbelMessages()));
     }
 
-    public void clearRBelMessages() {
+    public void clearRbelMessages() {
         LocalProxyRbelMessageListener.getValidatableRbelMessages().clear();
     }
 
@@ -111,19 +111,21 @@ public class RbelMessageValidator {
         }
     }
 
-    public void waitForMessageToBePresent(final RequestParameter requestParameter) {
-        findMessageByDescription(requestParameter);
+    public RbelElement waitForMessageToBePresent(final RequestParameter requestParameter) {
+        return findMessageByDescription(requestParameter);
     }
 
     protected RbelElement findMessageByDescription(final RequestParameter requestParameter) {
         final int waitsec = RBEL_REQUEST_TIMEOUT.getValue().orElse(5);
+
+        Optional<RbelElement> initialElement = getInitialElement(requestParameter);
 
         final AtomicReference<RbelElement> candidate = new AtomicReference<>();
         try {
             await("Waiting for matching request").atMost(waitsec, TimeUnit.SECONDS)
                 .pollDelay(0, TimeUnit.SECONDS).pollInterval(400, TimeUnit.MILLISECONDS)
                 .until(() -> {
-                    final Optional<RbelElement> found = filterRequests(requestParameter);
+                    final Optional<RbelElement> found = filterRequests(requestParameter, initialElement);
                     found.ifPresent(candidate::set);
                     return found.isPresent();
                 });
@@ -147,21 +149,33 @@ public class RbelMessageValidator {
         return candidate.get();
     }
 
-    protected Optional<RbelElement> filterRequests(final RequestParameter requestParameter) {
-
-        List<RbelElement> msgs = getRbelMessages();
+    private Optional<RbelElement> getInitialElement(RequestParameter requestParameter) {
         if (requestParameter.isStartFromLastRequest()) {
-            final RbelElement prevRequest = getCurrentRequest();
+            return LocalProxyRbelMessageListener.getValidatableRbelMessages().stream()
+                .dropWhile(msg -> msg != currentRequest)
+                .skip(1)
+                .findFirst();
+        } else if (requestParameter.isRequireNewMessage() && !LocalProxyRbelMessageListener.getValidatableRbelMessages().isEmpty()) {
+            return Optional.ofNullable(LocalProxyRbelMessageListener.getValidatableRbelMessages().getLast());
+        } else {
+            return Optional.empty();
+        }
+    }
+
+    protected Optional<RbelElement> filterRequests(final RequestParameter requestParameter, Optional<RbelElement> startFromMessageExclusively) {
+        List<RbelElement> msgs = getRbelMessages();
+        if (startFromMessageExclusively.isPresent()) {
             int idx = -1;
             for (var i = 0; i < msgs.size(); i++) {
-                if (msgs.get(i) == prevRequest) {
+                if (msgs.get(i) == startFromMessageExclusively.get()) {
                     idx = i;
                     break;
                 }
             }
-            msgs = new ArrayList<>(msgs.subList(idx + 2, msgs.size()));
+            if (idx > 0) {
+                msgs = new ArrayList<>(msgs.subList(idx + 1, msgs.size()));
+            }
         }
-
         final String hostFilter = TigerGlobalConfiguration.readString("tiger.rbel.request.filter.host", "");
         final String methodFilter = TigerGlobalConfiguration.readString("tiger.rbel.request.filter.method", "");
 
@@ -244,11 +258,14 @@ public class RbelMessageValidator {
         }
     }
 
-    public boolean doesHostMatch(final RbelElement req, final String hostFilter) {
+    public static boolean doesHostMatch(final RbelElement req, final String hostFilter) {
         try {
-            final String host = req.getFacetOrFail(RbelHttpMessageFacet.class)
-                .getHeader().getFacetOrFail(RbelHttpHeaderFacet.class)
-                .get("Host").getRawStringContent();
+            final String host = req.getFacet(RbelHttpMessageFacet.class)
+                .map(RbelHttpMessageFacet::getHeader)
+                .flatMap(el -> el.getFacet(RbelHttpHeaderFacet.class))
+                .map(el -> el.get("Host"))
+                .map(RbelElement::getRawStringContent)
+                .orElse("");
             return StringUtils.equals(host, hostFilter) || host.matches(hostFilter);
         } catch (final RuntimeException rte) {
             log.error("Probable error while parsing regex!", rte);
@@ -258,8 +275,11 @@ public class RbelMessageValidator {
 
     public boolean doesMethodMatch(final RbelElement req, final String method) {
         try {
-            final String reqMethod = req.getFacetOrFail(RbelHttpRequestFacet.class).getMethod().getRawStringContent()
-                .toUpperCase();
+            final String reqMethod = req.getFacet(RbelHttpRequestFacet.class)
+                .map(RbelHttpRequestFacet::getMethod)
+                .map(RbelElement::getRawStringContent)
+                .map(String::toUpperCase)
+                .orElse("");
             return method.equals(reqMethod) || method.matches(reqMethod);
         } catch (final RuntimeException rte) {
             log.error("Probable error while parsing regex!", rte);
@@ -503,7 +523,7 @@ public class RbelMessageValidator {
                 .flatMap(el -> el.getFacet(RbelValueFacet.class))
                 .map(RbelValueFacet::getValue)
                 .map(Object::toString)
-                .orElseThrow(() -> new NoSuchElementException("Unable to find a matching element for '" + rbelPath+ "'"));
+                .orElseThrow(() -> new NoSuchElementException("Unable to find a matching element for '" + rbelPath + "'"));
         }
     }
 }